Remove use of "visibility: collapsed" to hide filtered rows.
This is due to inconsistent browser support. For example Safari treated "visibility: collapsed" as "visibility: hidden", which
caused large blank tables. Now a height is set on the table container to hide filtered rows with overflow. "display: none" can't
be used as this would cause column widths to be recalculated.