Update generated table markup for the flat variant.
The previous markup was invalid. VoiceOver, and potentially other
assistive tech, would report an empty table.

To preserve table semantics for VoiceOver avoid setting the
`display` property on table elements. To achieve the visual look
of the flat table each row becomes a new table body with a row
for each data cell. A header "item x" is used to help seperate
the different rows to screen reader users. The `headers`
attribute is set to connect the headers to the correct data cells.
As per https://www.w3.org/WAI/tutorials/tables/multi-level/

Unfortunately HTML CodeStiffer doesn't approve of this approach,
shown in the WAI example. So we still have pa11y errors.