Remove flat table "item" headings.
HTML CodeSniffer doesn't approve of using the `headers` attribute
to link a heading which spans multiple columns to all following
rows, untill the next spanning header. Althought there is a
WAI example to this effect the experience isn't great in VoiceOver.
It did announce the heading cell if stepping through each cell
of the table but not whilst browsing following cells.
https://www.w3.org/WAI/tutorials/tables/multi-level/

The table markup is now valid on mobile devices; it is no longer
announced as a seperate table. There is room for improvement
regarding how each item in the table is delineated when flat.